Conflicts:
	.travis.yml
	cmake/util.cmake
	code/bmpman/bmpman.cpp
	code/cfile/cfile.cpp
	code/cfile/cfile.h
	code/cfile/cfilelist.cpp
	code/cfile/cfilesystem.cpp
	code/controlconfig/controlsconfigcommon.cpp
	code/ddsutils/ddsutils.cpp
	code/fs2netd/fs2netd_client.cpp
	code/globalincs/pstypes.h
	code/graphics/gropenglshader.cpp
	code/menuui/readyroom.cpp
	code/mission/missioncampaign.cpp
	code/mission/missionload.cpp
	code/network/multiutil.cpp
	code/parse/generic_log.cpp
	code/parse/lua.cpp
	code/parse/parselo.h
	code/pilotfile/csg.cpp
	code/pilotfile/csg_convert.cpp
	code/pilotfile/pilotfile_convert.cpp
	code/pilotfile/plr.cpp
	code/pilotfile/plr_convert.cpp
	code/playerman/managepilot.cpp
	code/sound/audiostr.cpp
	code/sound/audiostr.h
	code/sound/ogg/ogg.cpp
	lib/CMakeLists.txt

+++ code/bmpman/bm_internal.h
@@ -22,7 +22,6 @@
 #endif
 
 #include "bmpman/bmpman.h"
-#include "globalincs/pstypes.h"
 
 union bm_extra_info {
 	struct {
@@ -63,7 +62,7 @@ struct bitmap_entry {
 	cfile::DirType		dir_type; //!< which directory this was loaded from (to skip other locations with same name)
 
 	// compressed bitmap stuff (.dds) - RT please take a look at this and tell me if we really need it
-	int mem_taken;          //!< How much memory does this bitmap use? - UnknownPlayer
+	size_t mem_taken;          //!< How much memory does this bitmap use? - UnknownPlayer
 	int num_mipmaps;        //!< number of mipmap levels, we need to read all of them
 
 	// Stuff to keep track of usage
