+++ code/graphics/gropenglbmpman.cpp
@@ -9,21 +9,19 @@
 
 
 
-#include "globalincs/pstypes.h"
-#include "bmpman/bmpman.h"
-#include "graphics/gropenglbmpman.h"
-#include "ddsutils/ddsutils.h"
-#include "tgautils/tgautils.h"
-#include "pngutils/pngutils.h"
-#include "jpgutils/jpgutils.h"
-#include "pcxutils/pcxutils.h"
-#include "graphics/gropengltexture.h"
-#include "graphics/gropenglextension.h"
-#include "graphics/gropenglstate.h"
-#include "globalincs/systemvars.h"
 #include "anim/animplay.h"
 #include "anim/packunpack.h"
 #include "cmdline/cmdline.h"
+#include "ddsutils/ddsutils.h"
+#include "globalincs/systemvars.h"
+#include "graphics/gropenglbmpman.h"
+#include "graphics/gropenglextension.h"
+#include "graphics/gropenglstate.h"
+#include "graphics/gropengltexture.h"
+#include "jpgutils/jpgutils.h"
+#include "pcxutils/pcxutils.h"
+#include "pngutils/pngutils.h"
+#include "tgautils/tgautils.h"
 
 #define BMPMAN_INTERNAL
 #include "bmpman/bm_internal.h"
@@ -556,7 +554,7 @@ void gr_opengl_bm_save_render_target(int n)
 	bitmap_entry *be = &bm_bitmaps[n];
 	bitmap *bmp = &be->bm;
 
-	int rc = opengl_export_render_target( n, bmp->w, bmp->h, (bmp->true_bpp == 32), be->num_mipmaps, (ubyte*)bmp->data );
+	size_t rc = opengl_export_render_target( n, bmp->w, bmp->h, (bmp->true_bpp == 32), be->num_mipmaps, (ubyte*)bmp->data );
 
 	if (rc != be->mem_taken) {
 		Int3();
